fix:rename the mainbody component
