+++ src/components/index.ts
@@ -3,5 +3,5 @@ export * from './Header';
 export * from './SelectBox';
 export * from './languageSelectbox';
 export * from './SearchBar'
-export * from './MainBody'
-export * from './Greetings'
+export * from './PortfolioList'
+export * from './PortfolioHeader'
