+++ src/pages/Portfolios.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { MainBody, Header, Greetings } from "../components";
+import { Header, PortfolioList, PortfolioHeader } from "../components";
 
 export const Portfolio: React.FC = () => {
     const [searchValue, setSearchValue] = useState<string>("")
@@ -12,8 +12,8 @@ export const Portfolio: React.FC = () => {
             <Header />
             <main className="w-full h-screen overflow-y-auto pt-10 bg-gray-100">
                 <div className="px-12 pb-5 sm:px-24 md:px-28 lg:px-36">
-                    <Greetings onSearchChange={handleSearch} />
-                    <MainBody searchValue={searchValue} />
+                    <PortfolioHeader onSearchChange={handleSearch} />
+                    <PortfolioList searchValue={searchValue} />
                 </div>
             </main>
         </div>
