+++ src/types/type.ts
@@ -1,14 +1,3 @@
-import { ButtonType, ButtonVariant } from "./enum";
-
-export type ButtonProps = {
-    label: string;
-    style?: string;
-    type?: ButtonType
-    variant?: ButtonVariant;
-    isImg:boolean;
-    onClick: () => void;
-}
-
 export type CardTypes={
     image:string;
     title:string;
